tizzl/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
openai==1.35.3
anthropic==0.29.0
numpy==1.26.4
scikit-learn==1.5.0
chromadb==0.5.0
pillow==10.3.0
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
sentence-transformers==3.0.1
torch==2.3.1
torchvision==0.18.1
redis==5.0.4
boto3==1.34.122
pandas==2.2.2

# Testing dependencies
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0
faker==25.8.0